rename menu/index to menu/legacy
